+++ CHANGELOG.md
@@ -29,6 +29,8 @@
 * Rename `TrailingSemicolonAfterPropertyValue` to `TrailingSemicolon`
 * Teach `TrailingSemicolon` to report lints for uses of `@extend`, `@include`,
   or variable declarations without semicolons
+* Add `ignored_types` option to `CapitalizationInSelector` allowing you to selectively
+  ignore capitalization in certain types of selectors
 
 ## 0.25.1